LangGraph_poc/app3/pyproject.toml
[project]
name = "app3"
version = "0.1.0"
description = "Resume tailoring multi-agent app"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "langgraph>=1.1.6",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "reportlab>=4.0.0",
    "uvicorn>=0.44.0",
]

[tool.uv]
dev-dependencies = ["pytest>=8.0"]
